Don't add `accept-encoding` to grpc requests
