+++ tower-http/src/compression_utils.rs
@@ -3,7 +3,7 @@
 use bytes::{Bytes, BytesMut};
 use futures_core::Stream;
 use futures_util::ready;
-use http::HeaderValue;
+use http::{HeaderMap, HeaderValue};
 use http_body::Body;
 use pin_project::pin_project;
 use std::{
@@ -308,3 +308,27 @@ where
 }
 
 pub(crate) const SENTINEL_ERROR_CODE: i32 = -837459418;
+
+#[allow(clippy::clippy::needless_bool)]
+pub(crate) fn supports_transparent_compression(headers: &HeaderMap) -> bool {
+    let content_type = if let Some(content_type) = content_type(headers) {
+        content_type
+    } else {
+        return true;
+    };
+
+    if content_type == "application/grpc" {
+        // grpc doesn't support transparent compression and instead has its compression own
+        // algorithm that implementations can use
+        // https://grpc.github.io/grpc/core/md_doc_compression.html
+        false
+    } else {
+        // for now just say that all non-grpc requests support compression
+        true
+    }
+}
+
+fn content_type(headers: &HeaderMap) -> Option<&str> {
+    let content_type = headers.get(http::header::CONTENT_TYPE)?;
+    content_type.to_str().ok()
+}
